requirements.txt
torch>=2.0.0
transformers>=4.30.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.80.0
